Tests on ATMega328P, utility fix

+++ src/AvrSamples/Include/Utility.h
@@ -20,14 +20,13 @@ int IsNullOrEmpty(const char * str);
 int IntToString(long long int value, const char * format, char * buffer, int length);
 int UnsignedIntToString(unsigned long long int value, const char * format, char * buffer, int length);
 
-int ConvertToInt32(const char * str);
+long int ConvertToInt32(const char * str);
 long long int ConvertToInt(const char * str, const char * format);
 
 uint8_t ConvertToUInt8(const char * str);
 unsigned long long int ConvertToUnsignedInt32(const char * str);
 
 int FloatToString(long double value, const char * format, char * buffer, int length);
-
 float ConvertToFloat(const char * str);
 double ConvertToDouble(const char * str);
